dm-inject: Data structure corruption code implemented for superblock, group descriptor, bitmaps, inode table.
Known Issues:
submit_bio_wait does not work when called from ext4_*_from_dev.
this leads us to problems while processing directory structures, extent tree structures
and other such structures where we need to make a subsiquent call to bio for requesting
for a block that would be known to us on demand.